modularizando codigo base de websocket e organizando binaries
