+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cli-game-of-life"
+name = "gol"
 version = "0.1.0"
 edition = "2021"
