+++ src/client.rs
@@ -1,63 +1,19 @@
-use async_trait::async_trait;
 use ezsockets::ClientConfig;
+use gol::game::elements::point::Point;
+use gol::websocket::client::client::Client;
 use std::io::BufRead;
 
-#[derive(Debug)]
-struct Point {
-    x: usize,
-    y: usize,
-}
-
-impl From<&[u8]> for Point {
-    // add code here
-    fn from(bytes: &[u8]) -> Self {
-        let x = usize::from_le_bytes(bytes[0..8].try_into().expect("Slice with incorrect lenght"));
-        let y = usize::from_le_bytes(
-            bytes[8..16]
-                .try_into()
-                .expect("Slice with incorrect lenght"),
-        );
-        Point { x, y }
-    }
-}
-
-impl Into<Vec<u8>> for Point {
-    fn into(self) -> Vec<u8> {
-        let mut bytes = Vec::with_capacity(16);
-        bytes.extend_from_slice(&self.x.to_le_bytes());
-        bytes.extend_from_slice(&self.y.to_le_bytes());
-        bytes
-    }
-}
-
-struct Client {}
-
-#[async_trait]
-impl ezsockets::ClientExt for Client {
-    type Call = ();
-
-    async fn on_text(&mut self, text: String) -> Result<(), ezsockets::Error> {
-        tracing::info!("Mensagem recebida: {text}");
-        Ok(())
-    }
-
-    async fn on_binary(&mut self, bytes: Vec<u8>) -> Result<(), ezsockets::Error> {
-        let point: Point = From::from(&bytes[..]);
-
-        tracing::info!("{:?}", point);
-        Ok(())
-    }
-
-    async fn on_call(&mut self, _call: Self::Call) -> Result<(), ezsockets::Error> {
-        Ok(())
-    }
-}
-
 #[tokio::main]
-async fn main() {
+pub async fn main() {
     tracing_subscriber::fmt::init();
     let config = ClientConfig::new("ws://localhost:8080");
-    let (handle, future) = ezsockets::connect(|_client| Client {}, config).await;
+    let (handle, future) = ezsockets::connect(
+        |_client| Client {
+            p: Point::new(1, 2),
+        },
+        config,
+    )
+    .await;
     tokio::spawn(async move {
         future.await.unwrap();
     });
@@ -65,7 +21,6 @@ async fn main() {
     let stdin = std::io::stdin();
     let lines = stdin.lock().lines();
     for line in lines {
-        // Tente parsear a linha em dois valores usize
         if let Ok(pars) = line {
             let parts: Vec<&str> = pars.split_whitespace().collect();
 
@@ -77,7 +32,7 @@ async fn main() {
                 continue;
             }
 
-            let x: usize = match parts[0].parse() {
+            let row: usize = match parts[0].parse() {
                 Ok(val) => val,
                 Err(_) => {
                     tracing::error!("Falha ao parsear x: {}", parts[0]);
@@ -85,7 +40,7 @@ async fn main() {
                 }
             };
 
-            let y: usize = match parts[1].parse() {
+            let col: usize = match parts[1].parse() {
                 Ok(val) => val,
                 Err(_) => {
                     tracing::error!("Falha ao parsear y: {}", parts[1]);
@@ -93,8 +48,7 @@ async fn main() {
                 }
             };
 
-            // Crie a struct Point
-            let point = Point { x, y };
+            let point = Point::new(row, col);
 
             handle.binary(point).unwrap();
         }
