+++ src/lib.rs
@@ -1,3 +1,3 @@
 pub mod configuration;
 pub mod game;
-
+pub mod websocket;
